+++ include/Util.h
@@ -46,7 +46,7 @@ do {\
 	FUNC;\
 	}\
 	std::chrono::high_resolution_clock::time_point t2 = std::chrono::high_resolution_clock::now();\
-	std::chrono::duration<double> time_span = std::chrono::duration_cast<std::chrono::duration<double>>(t2 - t1);\
+	std::chrono::duration<float> time_span = std::chrono::duration_cast<std::chrono::duration<float>>(t2 - t1);\
 	std::cout << "TIMEIT ";\
 	std::cout.width(10);\
 	std::cout << std::left << time_span.count() << " s ";\
