+++ src/models/ComplexCell.cpp
@@ -34,7 +34,7 @@ void ComplexCell::doCalc(size_t cores, T* tca) {
 	// 2.1.	: UpdateVelocities
 	ca->calculatePaperMaxSpeed();
 	ca->adjustDt();
-	for (double t = 0.0; t < 1.0; t += ca->getDt())
+	for (float t = 0.0; t < 1.0; t += ca->getDt())
 	{
 		Cell_T::runAsync(cores, updateVelocities<T>, ca);
 		// 2.1.1 : EnforceBoundaryConditions
@@ -71,13 +71,13 @@ void ComplexCell::fixPaperHeight()
 	c = h * (C_MAX - C_MIN) + C_MIN;
 }
 
-void ComplexCell::setHeightGradient(double new_hx, double new_hy)
+void ComplexCell::setHeightGradient(float new_hx, float new_hy)
 {
 	hx = new_hx;
 	hy = new_hy;
 }
 
-void ComplexCell::addWater(double w)
+void ComplexCell::addWater(float w)
 {
 	u = fRand(3.0, 3.4);
 	v = fRand(-0.5, 0.8);
@@ -109,12 +109,12 @@ void ComplexCell::updateVelocities(size_t x, size_t y, T* ca)
 	auto* cr = ca->getOld(x + 1, y);
 	auto* cl = ca->getOld(x - 1, y);
 
-	double A = 0.0;
+	float A = 0.0;
 	A += pow((cc->u + cl->u) / 2.0, 2.0);
 	A -= pow((cc->u + cr->u) / 2.0, 2.0);
 	A += (cc->u + cd->u) * (cd->v + ca->getOld(x + 1, y - 1)->v) / 4.0;
 	A -= (cc->u + cu->u) * (cc->v + cr->v) / 4.0;
-	double B = 0.0;
+	float B = 0.0;
 	B += cr->u;
 	B += cl->u;
 	B += cu->u;
@@ -139,18 +139,18 @@ void ComplexCell::updateVelocities(size_t x, size_t y, T* ca)
 	ca->getNext(x, y)->v = cc->v + ca->getDt() * (A - WC_U * B + cc->p - cu->p - WC_K * cc->v);
 	/*
 	Paper* ca = static_cast<Paper*>(tca); // try fixing with 0.5index as "own" value
-	double A = 0.0;
+	float A = 0.0;
 	A += pow(ca->getOld(x, y)->u, 2);
 	A -= pow(cr->u, 2);
 	A += ((ca->getOld(x, y)->u + ca->getOld(x + 1, y - 1)->u) / 2.0) * ((ca->getOld(x, y)->v + ca->getOld(x + 1, y - 1)->v) / 2.0); // this line is fishy
 	A -= ((ca->getOld(x, y)->u + ca->getOld(x + 1, y + 1)->u) / 2.0) * ((ca->getOld(x, y)->v + ca->getOld(x + 1, y + 1)->v) / 2.0); // this line is fishy
-	double B = 0.0;
+	float B = 0.0;
 	B += ((cr->u + ca->getOld(x + 2, y)->u) / 2.0);
 	B += ((cl->u + ca->getOld(x, y)->u) / 2.0);
 	B += ((cu->u + ca->getOld(x + 1, y + 1)->u) / 2.0);
 	B += ((cd->u + ca->getOld(x + 1, y - 1)->u) / 2.0);
 	B -= (4.0 * ((ca->getOld(x, y)->u + cr->u) / 2.0));
-	double val = 0.0;
+	float val = 0.0;
 	val += ((ca->getOld(x, y)->u + cr->u) / 2.0);
 	val += ca->getDt() * (A - WC_U * B + ca->getOld(x, y)->p - cr->p - WC_K * ((ca->getOld(x, y)->u - cr->u) / 2.0));
 	//50% adds to my (x,y)->u, 50% adds to my next neighbor (x+1,y)->u
@@ -234,7 +234,7 @@ void ComplexCell::relaxDivergence(size_t x, size_t y, T* ca)
 	auto* n_cd = ca->getNext(x, y - 1) ? ca->getNext(x, y - 1) : n_cc;
 	auto* n_cl = ca->getNext(x - 1, y) ? ca->getNext(x - 1, y) : n_cc;
 	
-	double div = 0.0;
+	float div = 0.0;
 	div -= o_cc->u;
 	div += o_cl->u;
 	div -= o_cc->v;
@@ -261,7 +261,7 @@ const Color ComplexCell::draw(Color base) const
 {
 	if (p > 1)
 		std::cout << p << "\n";
-	return Color({ 1.0 - p, 0, 0, 1.0 });
+	return Color({ 1.0f - p, 0, 0, 1.0 });
 }
 
 //needed for proper linking
